content/media/test/test_bug448534.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=448534
-->

<head>
  <title>Test for Bug 448534</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=448535">Mozilla Bug 448534</a>
<pre id="test">
<script class="testbody" type="text/javascript">

var manager = new MediaTestManager;

function loaded(event) {
  var v = event.target;
  if (v._finished)
    return;
  v.play();
}

function started(event) {
  var v = event.target;
  if (v._finished)
    return;
  ok(!v.paused, "Video should not be paused while playing");
  v.parentNode.removeChild(v);
  v._played = true;
}

function stopped(event) {
  var v = event.target;
  if (v._finished)
    return;
  v._finished = true;
  ok(v.paused, "Video should be paused after removing from the Document");
  manager.finished(v.token);
}


function startTest(test, token) {
  var v = document.createElement('video');
  v.token = token;
  manager.started(token);
  v.src = test.name;
  v._played = false;
  v._finished = false;
  v.addEventListener("loadedmetadata", loaded, false);
  v.addEventListener("play", started, false);
  v.addEventListener("pause", stopped, false);
  document.body.appendChild(v);
}

manager.runTests(gSmallTests, startTest);

</script>
</pre>
</body>
</html>
